add beforeLoad tests
