+++ utam-compiler/src/test/java/utam/compiler/grammar/UtamMethod_ChainTests.java
@@ -9,11 +9,13 @@ package utam.compiler.grammar;
 
 import utam.compiler.helpers.TranslationContext;
 import utam.compiler.representation.ChainMethod;
+import utam.compiler.representation.PageObjectValidationTestHelper.MethodInfo;
 import utam.core.declarative.representation.PageObjectMethod;
 import utam.compiler.representation.PageObjectValidationTestHelper;
 import utam.core.framework.consumer.UtamError;
 import org.testng.annotations.Test;
 
+import static org.testng.Assert.assertThrows;
 import static utam.compiler.grammar.TestUtilities.*;
 import static utam.compiler.grammar.UtamMethod.*;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -186,5 +188,43 @@ public class UtamMethod_ChainTests {
     method.compose = new UtamMethodAction[0];
     UtamError e = expectThrows(UtamError.class, () -> method.getMethod(context));
     assertThat(e.getMessage(), containsString(getErr(ERR_METHOD_REDUNDANT_TYPE)));
+  }
+
+  @Test
+  public void testChainMethod() {
+    MethodInfo methodInfo1 = new MethodInfo("chainTest", "Type3");
+    methodInfo1.addCodeLine("this.getEl1().getEl3()");
+    TranslationContext context = new DeserializerUtilities().getContext("chainMethod");
+    PageObjectValidationTestHelper.validateMethod(context.getMethod("chainTest"), methodInfo1);
+
+    // test list
+    MethodInfo methodInfo2 = new MethodInfo("chainTestList", "List<Type3>");
+    methodInfo2.addCodeLine(
+        "this.getEl2Element().stream().flatMap(element -> element.getEl3().stream()).collect(Collectors.toList())");
+    PageObjectValidationTestHelper.validateMethod(context.getMethod("chainTestList"), methodInfo2);
+  }
+
+  @Test
+  public void testChainRequiredFields() {
+    String nameMissing = "{\"type\": \"clickable\"}";
+    assertThrows(
+        UtamError.class, () -> getDeserializedObject(nameMissing, UtamMethodChainLink.class));
+    String typeMissing = "{\"element\": \"element\"}";
+    assertThrows(
+        UtamError.class, () -> getDeserializedObject(typeMissing, UtamMethodChainLink.class));
+  }
+
+  /** Tests that an empty statement array throws the appropriate exception */
+  @Test
+  public void testChainEmptyStatementsArrayThrows() {
+    String json = "{\"name\": \"chainMethod\",\"chain\": []}";
+    UtamError e =
+        expectThrows(
+            UtamError.class,
+            () ->
+                getDeserializedObject(json, UtamMethod.class)
+                    .getChainMethod(getTestTranslationContext()));
+    assertThat(
+        e.getMessage(), containsString(String.format(ERR_METHOD_EMPTY_STATEMENTS, "chainMethod")));
   }
 }
