+++ utam-compiler/src/test/resources/beforeLoadMethod.json
@@ -1,0 +1,12 @@
+{
+  "beforeLoad": [
+    {
+      "element" : "root",
+      "apply" : "isPresent"
+    },
+    {
+      "element" : "root",
+      "apply": "getText"
+    }
+  ]
+}
